refactor: 1단계 리팩토링 및
