+++ src/main/java/jwp/controller/HomeController.java
@@ -15,7 +15,6 @@ public class HomeController extends HttpServlet {
 
     @Override
     protected void doGet(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
-//        req.setAttribute("users", MemoryUserRepository.getInstance().findAll());
         RequestDispatcher rd = req.getRequestDispatcher("/home.jsp");
         rd.forward(req,resp);
     }
