+++ src/main/java/jwp/controller/LoginController.java
@@ -11,21 +11,23 @@ import javax.servlet.http.HttpServletResponse;
 import javax.servlet.http.HttpSession;
 import java.io.IOException;
 
+import static jwp.util.UserSessionUtils.USER_SESSION_KEY;
+
 @WebServlet("/user/login")
 public class LoginController extends HttpServlet {
     @Override
     protected void doPost(HttpServletRequest req, HttpServletResponse resp) throws IOException {
+        String userId = req.getParameter("userId");
+        String password = req.getParameter("password");
         User loginUser = MemoryUserRepository.getInstance().findUserById(req.getParameter("userId"));
 
-        if (loginUser != null) {
+        if (loginUser != null && loginUser.isSameUser(userId, password)) {
             // 로그인 성공 시 세션 정보 저장
             HttpSession session = req.getSession();
-            session.setAttribute("user", loginUser);
-
+            session.setAttribute(USER_SESSION_KEY, loginUser);
             resp.sendRedirect("/");
-        } else {
-            resp.sendRedirect("/user/login_failed.jsp");
+            return;
         }
-
+        resp.sendRedirect("/user/login_failed.jsp");
     }
 }
