+++ src/main/java/jwp/controller/LogoutController.java
@@ -12,25 +12,17 @@ import javax.servlet.http.HttpServletResponse;
 import javax.servlet.http.HttpSession;
 import java.io.IOException;
 
+import static jwp.util.UserSessionUtils.USER_SESSION_KEY;
+
 @WebServlet("/user/logout")
 public class LogoutController extends HttpServlet {
-    @Override
-    protected void doPost(HttpServletRequest req, HttpServletResponse resp) throws IOException {
-        // 세션 데이터 삭제
-        HttpSession session = req.getSession();
-        session.removeAttribute("user");
-
-        resp.sendRedirect("/");
-    }
-
     @Override
     protected void doGet(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
         // URL에서 userId 파라미터 값 가져오기
         // 세션 데이터 삭제
         HttpSession session = req.getSession();
-        session.removeAttribute("user");
+        session.removeAttribute(USER_SESSION_KEY);
 
         resp.sendRedirect("/");
-//        rd.forward(req,resp);
     }
 }
