+++ src/main/java/jwp/controller/UpdateUserController.java
@@ -1,0 +1,27 @@
+package jwp.controller;
+
+import core.db.MemoryUserRepository;
+import jwp.model.User;
+
+import javax.servlet.ServletException;
+import javax.servlet.annotation.WebServlet;
+import javax.servlet.http.HttpServlet;
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
+import java.io.IOException;
+
+@WebServlet("/user/update")
+public class UpdateUserController extends HttpServlet {
+    @Override
+    protected void doPost(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
+        User updatedUser = new User(req.getParameter("userId"),
+                req.getParameter("password"),
+                req.getParameter("name"),
+                req.getParameter("email"));
+
+        MemoryUserRepository.getInstance().changeUserInfo(updatedUser);
+
+        resp.sendRedirect("/user/userList");
+
+    }
+}
